Attempt to handle "misclicks" better using Google Places (#3)
When the reverse geocoder does not match a locality, do a radius
search, spam the map with the locations that are returned, pan and
zoom, and finally let the user pick from the pins.

It is only marginally better, because it looks clumsy and does not
seem to always work. Doing a proper "every click matches" requires
advanced manipulation of GIS data, which is not worth the effort
given that we ultimately want a search-by-name rather than a
match-by-position.