Use the `el` from the Guide container for the Step views
This solves a problem that prevented the step1view from rendering
when #participate was active: the element in which the step views
should be rendered does exist when it is needed, but it cannot be
found in the DOM yet, perhaps because Google Maps is slow at updating
the map controls. In addition, the step view `.el` was set before it
could be resolved, anyway; however, it was not sufficient to call
`.setElement` at a later time.

Given that the step views use the same `el` as the guide container,
it may be best make the latter the owner of the former.