Automatically open the popup if there is a single result (#8)
Besides automatically triggering the click event, this commit contains
a workaround for a gotcha in the Google Maps API. When a Marker is
clicked with a mouse by a human user, the `event` argument of the
event handler is created automatically by the Google Maps API,
containing information such as the position of the Marker on the Map.
However, when triggering the click event programmatically, the caller
has to provide this information by themselves. While it is possible to
do so, it is easier to change the handler so that it does not depend
on the `event` argument.