Include the country in the query, ditch the bounds
Fixes #25, fixes #26. The former use of the bounds property should be
correct according to the Google Maps documentation, but the GMaps API
doesn't hold up to its promise. So we circumvent the problem and
instead exploit the fact that the Places.textSearch API supports
complex queries.

With partial city names, search results can be a bit surprising. For
example, if you search for "san, AR", Google Places thinks that Sanaa
in Yemen is the closest match. This could possibly be improved upon
further by using the full name of the country, such as
"san, Argentina".