Make POST to /api/recordings/ CSRF-exempt
Obviously, this is less than ideal, but it is the fastest way to make
the uploading work. This should be replaced by a safe solution ASAP,
probably by deferring to the Django backend for the initial page load
and decorating the view in question with @ensure_csrf_cookie.

Strongly related to #58.