Rename the Recording model to Contribution
This turns out to be a dead end. Due to a bug in Django, applying the
migration leads to the following cryptical prompt:

    The following content types are stale and need to be deleted:

        recordings | recording

    Any objects related to these content types by a foreign key will also
    be deleted. Are you sure you want to delete these content types?
    If you're unsure, answer 'no'.

Answering 'yes' will cause data loss, but answering 'no' might leave
the database in an invalid state and it is difficult to tell whether
this is the case.

Committing this for future reference. Related to #57.