Auto-localize Google Maps
According to the documentation, Google Maps should base the
localization on the visitor computer's system settings when the
language and the region are not explicitly set. I was not able to
confirm this locally, but since I was always shown Dutch localization,
at least we can be sure it doesn't default to English.

docs: https://developers.google.com/maps/documentation/javascript/localization