Fix some bugs
The migration was missing a recording.save. Just reading the extension
from the file name turned out to be more reliable in the post-save
hook than mimetypes.guess_type; it may return 'audio/mpeg' for .m4a in
some cases, in which case the recording won't be converted to MP3.