Localize the jquery.validate messages (close #88)
This is a change from @oktaal's approach in 39c76f2-5bc9576.

Disadvantage of this new approach: some of jquery.validation's
additional methods lack localization, so their messages always appear
in English. In practice this is a minor concern, because it only
affects the requirement to fill at least one contact detail field and
this requirement is already documented in the label.

Advantage of the new approach: we don't copy the localization strings
from jquery.validate into our own project, so we don't need to
manually update them when they change, either. Less maintenance!